check is it perfect or not

+++ 16-binary_tree_is_perfect.c
@@ -6,7 +6,7 @@
 */
 int binary_tree_is_perfect(const binary_tree_t *tree)
 {
-	size_t right_height, left_height;
+	int right_height, left_height;
 
 	if (tree == NULL)
 	{
